CORE-2: Refactored Commands (CoreCommands can now be executed by Players & Console)
